Extract CSharp type body scope partial
